# Fix ambiguous conversion from long to QVariant
<==== Mercurial data ====>
Branch: vms

+++ mediaserver_core/src/analytics/events_storage/analytics_events_storage.cpp
@@ -127,7 +127,7 @@ void EventsStorage::insertEventAttributes(
         INSERT INTO event_properties(docid, content)
         VALUES(:eventId, :content)
     )sql");
-    insertEventAttributesQuery.bindValue(":eventId", eventId);
+    insertEventAttributesQuery.bindValue(":eventId", static_cast<qint64>(eventId));
     insertEventAttributesQuery.bindValue(
         ":content",
         containerString(eventAttributes, "; ", "", "", ""));
